Revert "LPD-52709 Create a JEE filter to handle frontend resources"
This reverts commit c8d41ac50d75390c463459cb3b6fd74822b0ae00.
