+++ modules/apps/frontend-js/frontend-js-web/src/main/java/com/liferay/frontend/js/web/internal/js/importmaps/extender/FrontendJsWebDynamicJSImportMapsContributor.java
@@ -15,7 +15,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.io.Writer;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -32,28 +35,23 @@ public class FrontendJsWebDynamicJSImportMapsContributor
 
 		writer.write("\"@liferay/language/\": \"/o/js/language/\"");
 
-		HashedFilesRegistry hashedFilesRegistry =
-			HashedFilesRegistry.getHashedFilesRegistry();
+		_hashedFilesRegistry.forEach(
+			(unhashedFileURI, hashedFileURI) -> {
+				if (!unhashedFileURI.endsWith(".js")) {
+					return;
+				}
 
-		if (hashedFilesRegistry != null) {
-			hashedFilesRegistry.forEach(
-				(unhashedFileURI, hashedFileURI) -> {
-					if (!unhashedFileURI.endsWith(".js")) {
-						return;
-					}
-
-					try {
-						writer.write(", \"");
-						writer.write(unhashedFileURI);
-						writer.write("\": \"");
-						writer.write(hashedFileURI);
-						writer.write(StringPool.QUOTE);
-					}
-					catch (IOException ioException) {
-						throw new RuntimeException(ioException);
-					}
-				});
-		}
+				try {
+					writer.write(", \"");
+					writer.write(unhashedFileURI);
+					writer.write("\": \"");
+					writer.write(hashedFileURI);
+					writer.write(StringPool.QUOTE);
+				}
+				catch (IOException ioException) {
+					throw new RuntimeException(ioException);
+				}
+			});
 	}
 
 	@Override
@@ -61,6 +59,20 @@ public class FrontendJsWebDynamicJSImportMapsContributor
 			HttpServletRequest httpServletRequest, Writer writer)
 		throws IOException {
 	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_hashedFilesRegistry = new HashedFilesRegistry(bundleContext);
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_hashedFilesRegistry.close();
+
+		_hashedFilesRegistry = null;
+	}
+
+	private HashedFilesRegistry _hashedFilesRegistry;
 
 	@Reference
 	private Portal _portal;
